StudentGradebookTests/SemesterControlSchedulesControllerTests.cs
using Xunit;
using Moq;
using Microsoft.EntityFrameworkCore;
using StudentGradebookWebAPI.Controllers;
using StudentGradebookWebAPI.Database;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Mvc;

public class SemesterControlSchedulesControllerTests
{
    private readonly Mock<StudentGradebookContext> _mockContext;
    private readonly SemesterControlSchedulesController _controller;

    public SemesterControlSchedulesControllerTests()
    {
        var mockSet = new Mock<DbSet<SemesterControlSchedule>>();
        _mockContext = new Mock<StudentGradebookContext>();
        _mockContext.Setup(m => m.SemesterControlSchedules).Returns(mockSet.Object);
        _controller = new SemesterControlSchedulesController(_mockContext.Object);
    }

    [Fact]
    public async Task GetSemesterControlSchedules_ReturnsAllSchedules()
    {
        // Arrange
        var mockData = new List<SemesterControlSchedule> { new SemesterControlSchedule(), new SemesterControlSchedule() };

        // Act

        // Assert
        Assert.Equal(true, true);
    }

    [Fact]
    public async Task GetSemesterScheduleForStudent_ReturnsCorrectSchedules()
    {
        // Arrange
        int studentId = 1;
        var mockData = new List<SemesterControlSchedule>
        {
            new SemesterControlSchedule { /* Initialize properties */ },
            new SemesterControlSchedule { /* Initialize properties */ }
        };


        // Assert
        Assert.Equal(true, true);
    }

    [Fact]
    public async Task GetSemesterControlSchedule_ReturnsScheduleById()
    {
        // Arrange
        int id = 1;
        var mockItem = new SemesterControlSchedule { EntryId = id };
        _mockContext.Setup(m => m.SemesterControlSchedules.FindAsync(id)).ReturnsAsync(mockItem);

        // Act
        var result = await _controller.GetSemesterControlSchedule(id);

        // Assert
        var actionResult = Assert.IsType<ActionResult<SemesterControlSchedule>>(result);
        var returnValue = Assert.IsType<SemesterControlSchedule>(actionResult.Value);
        Assert.Equal(id, returnValue.EntryId);
    }
}
